ADM-972 [frontend]: refactor
